change type of node-state for slot

+++ MasterNodeState.sol
@@ -6,22 +6,19 @@ import "./System.sol";
 contract MasterNodeState is INodeState, System {
     uint[] ids; // all node id
     mapping(uint => uint) id2index; // index in ids
-    mapping(uint => uint8) id2state; // id to state
+    mapping(uint => uint) id2state; // id to state
     mapping(uint => StateEntry[]) id2entries; // id to upload informations
 
-    function uploadState(uint[] memory _ids, uint8[] memory _states) public onlySN {
+    function uploadState(uint[] memory _ids, uint[] memory _states) public onlySN {
         require(_ids.length == _states.length, "id list isn't matched with state list");
-
         bool flag = false;
         uint pos = 0;
-
         for(uint i = 0; i < _ids.length; i++) {
             uint id = _ids[i];
             if(!getMasterNode().existID(id)) {
                 continue;
             }
-
-            uint8 state = _states[i];
+            uint state = _states[i];
             flag = false;
             pos = 0;
             (flag, pos) = existCaller(id, msg.sender);
@@ -30,7 +27,6 @@ contract MasterNodeState is INodeState, System {
             } else {
                 id2entries[id].push(StateEntry(msg.sender, state));
             }
-
             updateState(id, state);
         }
     }
@@ -58,7 +54,7 @@ contract MasterNodeState is INodeState, System {
         return (false, 0);
     }
 
-    function updateState(uint _id, uint8 _state) internal {
+    function updateState(uint _id, uint _state) internal {
         StateEntry[] storage entries = id2entries[_id];
         uint num = 0;
         for(uint i = 0; i < entries.length; i++) {
